requirements.txt
flask
flask-wtf
flask-login
mysql-connector-python